fix(tour): demo拆解与规范
